requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0

# Data Acquisition
fredapi>=0.5.2
yfinance>=0.2.66
python-dotenv>=1.0.0

# Modeling
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Explainability
shap>=0.44.0

# Application
streamlit>=1.30.0
plotly>=5.0.0
matplotlib>=3.7.0